+++ landingzones/tests/test_views_taskflow.py
@@ -8,6 +8,7 @@ from irods.test.helpers import make_object
 from irods.keywords import REG_CHKSUM_KW
 
 from django.contrib import auth
+from django.contrib.messages import get_messages
 from django.core import mail
 from django.test import override_settings
 from django.urls import reverse
@@ -32,6 +33,7 @@ from timeline.models import ProjectEvent
 
 from landingzones.models import LandingZone
 from landingzones.tests.test_models import LandingZoneMixin
+from landingzones.views import ZONE_MOVE_INVALID_STATUS
 
 
 User = auth.get_user_model()
@@ -333,13 +335,11 @@ class TestLandingZoneMoveView(
 
     def setUp(self):
         super().setUp()
-
         # Get iRODS backend for session access
         self.irods_backend = get_backend_api('omics_irods')
         self.assertIsNotNone(self.irods_backend)
         self.irods_session = self.irods_backend.get_session()
 
-        # Init project
         # Make project with owner in Taskflow and Django
         self.project, self.owner_as = self.make_project_taskflow(
             title='TestProject',
@@ -366,7 +366,6 @@ class TestLandingZoneMoveView(
             configuration=None,
             config_data={},
         )
-
         # Create zone in taskflow
         self.make_zone_taskflow(self.landing_zone)
         # Get collections
@@ -597,6 +596,52 @@ class TestLandingZoneMoveView(
         self.assertTrue('BatchCheckFilesTask' in zone.status_info)
         self.assertEqual(len(self.zone_coll.data_objects), 1)
         self.assertEqual(len(self.assay_coll.data_objects), 0)
+
+    def test_move_invalid_status(self):
+        """Test validating and moving with invalid zone status (should fail)"""
+        self.irods_obj = self.make_object(self.zone_coll, TEST_OBJ_NAME)
+        self.md5_obj = self.make_md5_object(self.irods_obj)
+        zone = LandingZone.objects.first()
+        zone.status = 'VALIDATING'
+        zone.save()
+        self.assertEqual(len(self.zone_coll.data_objects), 2)
+        self.assertEqual(len(self.assay_coll.data_objects), 0)
+        self.assertEqual(len(mail.outbox), 1)
+        self.assertEqual(
+            ProjectEvent.objects.filter(event_name='zone_move').count(), 0
+        )
+        self.assertEqual(
+            AppAlert.objects.filter(alert_name='zone_move').count(), 0
+        )
+
+        with self.login(self.user):
+            response = self.client.post(
+                reverse(
+                    'landingzones:move',
+                    kwargs={'landingzone': self.landing_zone.sodar_uuid},
+                ),
+            )
+            self.assertRedirects(
+                response,
+                reverse(
+                    'landingzones:list',
+                    kwargs={'project': self.project.sodar_uuid},
+                ),
+            )
+
+        self.assertEqual(
+            str(list(get_messages(response.wsgi_request))[0]),
+            ZONE_MOVE_INVALID_STATUS,
+        )
+        self.assert_zone_status(zone, 'VALIDATING')
+        self.assertEqual(len(self.zone_coll.data_objects), 2)
+        self.assertEqual(len(self.assay_coll.data_objects), 0)
+        self.assertEqual(len(mail.outbox), 1)
+        tl_event = ProjectEvent.objects.filter(event_name='zone_move').first()
+        self.assertIsNone(tl_event)
+        self.assertEqual(
+            AppAlert.objects.filter(alert_name='zone_move').count(), 0
+        )
 
     @override_settings(REDIS_URL=INVALID_REDIS_URL)
     def test_move_lock_failure(self):